Refactor qualities selector to client component
